fix: the Boat path was dead on arrival, and my own fix made a host look infinite
Two criticals from the review, both invisible to a green suite.

**cpu_max_cores crossed as a float into an int, so every PUT was a 400.**
Atlas sells fractional CPU — a Shared 1x VM gets a quarter of a core — so
its field is a float, and Python's encoder always writes a decimal point,
which Go refuses to decode into an integer even for 1.0. WO-2 put a PUT
in front of every verb, so the entire Boat path failed before it did
anything, for every VM on every host. It survived because the spec claims
both clients are generated from the IDL and only the Go one is: nothing
in sixty Atlas tests validates a request body against the contract. The
contract now types it a number, which is what it always was.

**An unmeasured fact must be absent, not zero — my earlier fix had this
backwards.** I tolerated an unreadable thin pool and reported 0,
reasoning that Atlas treats zero as "unmeasured, leave what you had". It
does not: `server_capacity.py` reads a falsy total as None, and None
means UNLIMITED on that axis, so placement stops gating disk entirely. A
host whose pool Boat cannot read would have kept accepting VMs forever.
That is strictly worse than the 500 it replaced. The facts are optional
on the wire now, so the mirror's is-not-None guard can actually fire.

Also: a failed fullness read no longer discards a size that WAS measured
— the size is the placement input and the fullness is only an alert — and
an absent Firecracker no longer overwrites the version the Server row
knew.

+++ internal/wire/wire.gen.go
@@ -71,10 +71,17 @@ type DesiredVirtualMachine struct {
 	// Epochs start at 1. Boat gates on whether it holds a fence at all, so
 	// a zero would be a held fence rather than the absence of one — do not
 	// use 0 to mean unfenced.
-	BootEpoch         int64   `json:"boot_epoch"`
-	CpuMaxCores       *int    `json:"cpu_max_cores,omitempty"`
-	CpuMode           *string `json:"cpu_mode,omitempty"`
-	DataDiskGigabytes *int    `json:"data_disk_gigabytes,omitempty"`
+	BootEpoch int64 `json:"boot_epoch"`
+
+	// CpuMaxCores A number, not an integer: a share of a core is a real product. Atlas
+	// sells fractional CPU (a Shared 1x VM gets a quarter of a core), its
+	// own field is a float, and Python's JSON encoder always writes a
+	// float with a decimal point — which Go refuses to decode into an
+	// integer, even for 1.0. Typed as an integer, this rejected every
+	// desired-state assertion Atlas could send, for every VM.
+	CpuMaxCores       *float32 `json:"cpu_max_cores,omitempty"`
+	CpuMode           *string  `json:"cpu_mode,omitempty"`
+	DataDiskGigabytes *int     `json:"data_disk_gigabytes,omitempty"`
 
 	// DesiredPower The only power input Boat's reconciler takes. An explicit Stopped
 	// outranks a wake trap: a VM that was told to stop is not woken by
